refactor(downloader): move DownloadStatus enumeration from datasource to downloader constants
