--- src/GOES_DL/datasource/constants.py
@@ -1,32 +1,0 @@
-"""
-Define and export constants used across the GOES-DL datasource module.
-
-These constants include date and time formats, and other fixed values
-that are used throughout the module.
-
-Classes
--------
-DownloadStatus:
-    Enumeration for download status.
-"""
-
-from enum import Enum
-
-
-class DownloadStatus(Enum):
-    """
-    Enumeration for download status.
-
-    Attributes
-    ----------
-    SUCCESS
-        Indicates that the download was successful.
-    FAILED
-        Indicates that the download failed.
-    ALREADY
-        Indicates that the file was already downloaded.
-    """
-
-    SUCCESS = 0
-    FAILED = 1
-    ALREADY = 2
